#7 - adjustedg about file'name and SSR in file page.tsx
