+++ client/src/app/authentic/page.tsx
@@ -1,17 +1,17 @@
-"use client";
+// "use client";
 
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
+// import { useEffect } from "react";
+// import { useRouter } from "next/navigation";
 
-export default function AuthPage() {
-  const router = useRouter();
+// export default function AuthPage() {
+//   const router = useRouter();
 
-  useEffect(() => {
-    router.replace("/authentic/signIn"); // Chuyển hướng về trang đăng nhập
-  }, [router]);
+//   useEffect(() => {
+//     router.replace("/authentic/signIn"); // Chuyển hướng về trang đăng nhập
+//   }, [router]);
 
-  return null; // Không hiển thị gì cả, chỉ để chuyển hướng
-}
+//   return null; // Không hiển thị gì cả, chỉ để chuyển hướng
+// }
