+++ client/src/app/authentic/sent_email/page.tsx
@@ -1,8 +1,8 @@
-"use client";
+
 
 import dynamic from "next/dynamic";
 
-const Sent_mail = dynamic(() => import("~/components/Forgot Password_tsx/EmailSent_tsx/sent_email"));
+const Sent_mail = dynamic(() => import("~/components/ForgotPassword_tsx/EmailSent"));
 
 export default function Sent_page (){
     return (
