+++ client/src/app/authentic/signIn/page.tsx
@@ -1,8 +1,7 @@
-"use client";
 
 import dynamic from "next/dynamic";
 
-const SignInForm = dynamic(() => import("~/components/SignInForm_tsx/signInForm"));
+const SignInForm = dynamic(() => import("~/components/SignInForm_tsx"));
 
 export default function SignInPage() {
   return (
